+++ lib/core/card-view/components/card-view-textitem/card-view-textitem.component.spec.ts
@@ -22,7 +22,10 @@ import { CardViewUpdateService } from '../../services/card-view-update.service';
 import { CardViewTextItemComponent } from './card-view-textitem.component';
 import { setupTestBed } from '../../../testing/setup-test-bed';
 import { CoreTestingModule } from '../../../testing/core.testing.module';
-import { CardViewItemFloatValidator, CardViewItemIntValidator, CardViewIntItemModel, CardViewFloatItemModel } from '@alfresco/adf-core';
+import { CardViewItemFloatValidator } from '../../validators/card-view-item-float.validator';
+import { CardViewItemIntValidator} from '../../validators/card-view-item-int.validator';
+import { CardViewIntItemModel } from '../../models/card-view-intitem.model';
+import { CardViewFloatItemModel } from '../../models/card-view-floatitem.model';
 import { MatChipsModule } from '@angular/material/chips';
 import { ClipboardService } from '../../../clipboard/clipboard.service';
 import { DebugElement, SimpleChange } from '@angular/core';
