+++ lib/core/directives/logout.directive.spec.ts
@@ -20,7 +20,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { Router } from '@angular/router';
 import { of, throwError } from 'rxjs';
 import { AuthenticationService } from '../services';
-import { AppConfigService } from '../app-config';
+import { AppConfigService } from '../app-config/app-config.service';
 import { setupTestBed } from '../testing/setup-test-bed';
 import { LogoutDirective } from './logout.directive';
 import { CoreTestingModule } from '../testing/core.testing.module';
