+++ lib/core/services/identity-role.service.spec.ts
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import { setupTestBed } from '@alfresco/adf-core';
+import { setupTestBed } from '../testing/setup-test-bed';
 import { HttpClientModule, HttpErrorResponse, HttpResponse } from '@angular/common/http';
 import { TestBed } from '@angular/core/testing';
 import { of, throwError } from 'rxjs';
